+++ autotest/src/main/java/com/autotest/runner/DeviceSelector.kt
@@ -3,11 +3,6 @@ package com.autotest.runner
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.uiautomator.UiDevice
 
-enum class DevicePreference {
-    FIRST_AVAILABLE,
-    SPECIFIC_SERIAL
-}
-
 object DeviceSelector {
 
     fun getDevice(): UiDevice {